Additional parameters added
Mime types added
Simple validation added

+++ src/index.js
@@ -19,10 +19,10 @@ module.exports = co.wrap(function* (params) {
     .filter(Boolean);
 
   if (!fileNames.length) {
-    console.log('...Complete. Found no files to process. Exit');
+    console.log('Complete. Found no files to process. Exit');
     return null;
   }
-  console.log('...Complete. Found', fileNames.length, 'suitable files and folders');
+  console.log('Complete');
 
   const awsOptions = {
     sslEnabled: true,
@@ -38,40 +38,47 @@ module.exports = co.wrap(function* (params) {
   console.log('Computing map of hashes for local files');
   const localHashesMap = yield computeLocalFilesStats(fileNames, basePath, opts.concurrency);
   const localFilesAmount = Object.keys(localHashesMap).length;
-  console.log('...Complete. Found', localFilesAmount, 'files locally');
+  console.log('Complete. Found', localFilesAmount, 'files locally');
 
   console.log('Computing map of hashes for S3-stored files');
   const remoteHashesMap = yield s3HelperInstance.getRemoteFilesStats();
-  const remoteFilesAmount = Object.keys(localHashesMap).length;
-  console.log('...Complete. Found', remoteFilesAmount, 'files in S3');
+  const remoteFilesAmount = Object.keys(remoteHashesMap).length;
+  console.log('Complete. Found', remoteFilesAmount, 'files in S3');
 
   console.log('Computing difference');
   const { toUpload, toDelete } = detectFileChanges(localHashesMap, remoteHashesMap);
 
   const uploadNeeded = Object.keys(toUpload).length;
   if (uploadNeeded) {
-    console.log('%s files to be uploaded', uploadNeeded);
-    console.log('Uploading files');
+    console.log('%s files to be uploaded. Uploading files', uploadNeeded);
     yield s3HelperInstance.uploadObjects(toUpload, basePath);
-    console.log('...Complete');
+    console.log('Complete');
+  } else {
+    console.log('No files to be uploaded');
   }
 
-  const removalNeeded = Object.keys(toDelete).length;
-  if (removalNeeded) {
-    console.log('%s files to be removed', removalNeeded);
-    console.log('Removing files');
-    yield s3HelperInstance.deleteObjects(toDelete);
-    console.log('...Complete');
+  if (!opts.noRm) {
+    const removalNeeded = Object.keys(toDelete).length;
+    if (removalNeeded) {
+      console.log('%s files to be removed. Removing files', removalNeeded);
+      yield s3HelperInstance.deleteObjects(toDelete);
+      console.log('Complete');
+    } else {
+      console.log('No files to be removed');
+    }
+  } else {
+    console.log('Skipping removal as correspondent flag is set');
+    Object.assign(localHashesMap, toDelete);
   }
 
   console.log('Saving map of file hashes');
   yield s3HelperInstance.storeRemoteHashesMap(localHashesMap);
-  console.log('...Complete');
+  console.log('Complete');
 
   if (opts.cfDistId) {
     console.log('Creating CloudFront invalidation for', opts.cfDistId);
     const invalidateResponse = yield invalidate(new aws.CloudFront(), opts.cfDistId, opts.cdInvPaths);
     const invalidationId = invalidateResponse.Invalidation.Id;
-    console.log('...Complete. CloudFront invalidation created:', invalidationId);
+    console.log('Complete. CloudFront invalidation created:', invalidationId);
   }
 });
